修复wxuin的int类型导致编译的32位程序出错

+++ user.go
@@ -11,7 +11,7 @@ import (
 
 // User 抽象的用户结构: 好友 群组 公众号
 type User struct {
-	Uin               int
+	Uin               int64
 	HideInputBarFlag  int
 	StarFriend        int
 	Sex               int
